Move Rs2Option into it's own file / submodule
